modularizar el codigo
