+++ index.html
@@ -17,9 +17,6 @@
       background-color: #212529;
       color: #f8f9fa;
     }
-    /* .card { */
-    /*   margin-top: 20px; */
-    /* } */
     .list-group {
       height: 500px;
     }
@@ -69,6 +66,6 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="./script.js"></script>
+  <script type="module" src="./js/main.js"></script>
 </body>
 </html>
