+++ js/constants.js
@@ -1,0 +1,13 @@
+export const DARK_MODE = "dark-mode";
+export const LIGHT_MODE = "light-mode";
+export const THEME_KEY = "theme";
+export const DATA_KEY = "contactos";
+
+export const DOM = {
+  bodyEl: document.body,
+  navbar: document.querySelector(".navbar"),
+  themeToggleBtn: document.querySelector("#toggle-theme"),
+  formCard: document.querySelector("#form-card"),
+  form: document.querySelector("#agenda-form"),
+  contactList: document.querySelector("#contactos"),
+};
